server/requirements.txt
asgiref==3.7.2
Django==4.2.2
django-cors-headers==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.7.0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2023.3
sqlparse==0.4.4
Pillow==10.3.
requests==2.32.3
setuptools==70.0.0
web3==7.2.0
